+++ man-roxygen/return_cum_accumulate.R
@@ -1,0 +1,3 @@
+#' @return [`arrayish`]\cr
+#'   Has the same shape as the input, and the same data type, except for a
+#'   boolean input, which is accumulated at `i32`.