+++ man-roxygen/return_reduce.R
@@ -1,3 +1,3 @@
 #' @return [`arrayish`]\cr
-#'   Has <%= dtype_out %>, and the input's shape with the reduced axes removed
-#'   (`drop = TRUE`) or set to 1 (`drop = FALSE`).
+#'   Has <%= dtype_out %>. The shape is the input's with the reduced axes
+#'   removed (`drop = TRUE`) or set to 1 (`drop = FALSE`).